Move Ubisoft logic to external class
